% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcfunpca.R
\name{calcfunpca}
\alias{calcfunpca}
\title{Do dimensional reduction using functional pca.}
\usage{
calcfunpca(cross, pheno.cols, n.max = 4, criteria = 0.9, nbasis, nn = 0)
}
\arguments{
\item{cross}{An object of class \code{"cross"}. See the \code{\link[qtl]{read.cross}} for details.}

\item{pheno.cols}{Columns in the phenotype matrix to be used as the
phenotype.}

\item{n.max}{The number of maximum reduced dimension.}

\item{criteria}{how much of variance explained.}

\item{nbasis}{The number of basis to use.}

\item{nn}{The number of exact reduced dimension}
}
\value{
It gives a list, Y is a matrix that each column have principal components. eigf is a eigen function object from functional PCA using fda package.
}
\description{
Do dimensional reduction using functional pca.
}
\examples{
data(exd)
exd <- calc.genoprob(exd, step=2)
cvout <- cvfold(exd, basisset = 4:7, fold = 10)
cvout # basis number 5 have the smallest sse. So we take nbasis = 5.
Y <- calcfunpca(exd, criteria=0.9, nbasis = 5)$Y
out1 <- scanoneM(exd, Y, method = "hk")
}
\seealso{
\code{\link{scanoneM}}
}
\author{
Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\keyword{utilities}
