\name{exd}
\alias{exd}
\docType{data}

\title{An example data set}

\description{
  Simulated data for an backcross with a QTL at 1@42 with function
  valued phenotype.
}

\usage{data(exd)}

\format{
  An object of class \code{cross}.  See \code{\link{read.cross}}
  for details.
}

\details{
  There are 100 BC individuals typed at a total of 18 markers on three
  chromosomes.  The data were simulated with QTLs at the 42cM of
  chromosomes 1.

  The function valued phenotype is measured in 10 equilly spaced time points.
}

\examples{
data(exd)
}

\author{Il-Youp Kwak, \email{ikwak2@stat.wisc.edu} }

\keyword{datasets}