% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geteffects.R
\name{geteffects}
\alias{geteffects}
\title{Estimate QTL effects}
\usage{
geteffects(cross, pheno.cols)
}
\arguments{
\item{cross}{An object of class 'cross'. See 'read.cross' for details.}

\item{pheno.cols}{phenotype columns to be used}
}
\value{
A matrix of coefficients. (i,j)th item is a coefficient of jth
position as a qtl of ith observation.
}
\description{
Estimate QTL effect at each genomic position for each of multiple phenotypes.
However, this function uses a simple regression based method to predict the effect. It is recommended to use the fitqtl function referring to a package vignette to predict the effects of QTL.
}
\examples{
data(simspal)
simspal <- calc.genoprob(simspal)
phe <- 1:nphe(simspal)
\dontshow{phe <- seq(1, nphe(simspal), by=60)}
eff <- geteffects(simspal, pheno.cols=phe)

}
\seealso{
\code{\link{plotlod}}
}
\author{
Karl W Broman, Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\keyword{models}
