% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanoneF.R
\name{scanoneF}
\alias{scanoneF}
\title{Genome scan with a single QTL model}
\usage{
scanoneF(cross, pheno.cols, n.perm, ...)
}
\arguments{
\item{cross}{An object of class \code{"cross"}. See \code{\link[qtl]{read.cross}} for details.}

\item{pheno.cols}{Columns in the phenotype matrix to be used as the
phenotype.}

\item{n.perm}{If specified, a permutation test is performed rather than an
analysis of the observed data.  This argument defines the number of
permutation replicates.}

\item{\dots}{More parameters controled in 'scanone'. See 'scanone' for
details.}
}
\value{
If \code{n.perm} is missing, the function returns a data.frame whose
first two columns contain the chromosome IDs and cM positions.  Subsequent
third and fourth columns contain the SLOD and MLOD scores.

If \code{n.perm} is specified, the function returns the results of a permutation
test and the output returns the matrix of two columns. The first column for
SLOD and the second column for MLOD score.
}
\description{
Extension of the R/qtl function \code{\link[qtl]{scanone}}. Genome scan with a single
QTL model, with possible allowance for covariates, using several possible
models for the function valued phenotype.
}
\examples{
data(exd)

# calculate QTL genotype probabilities and perform genome scan
exd <- calc.genoprob(exd, step=2)
out <- scanoneF(exd)

# summarize results
summary(out)

# Plot the results : red for slod, blue for mlod
plot(out, lod=1:2, col = c("red","blue"), ylab = "lod")


# Permutation tests
nperm <- 1000
\dontshow{n.perm <- 5}
perm1 <- scanoneF(exd, method="hk", n.perm=n.perm)
summary(perm1, alpha=c(0.05, 0.10))

# Results above the 0.05 threshold
summary(out, perms=perm1, alpha=0.05)
}
\author{
Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\seealso{
\code{\link[qtl]{scanone}}, \code{\link{scanoneM}}
}
\keyword{models}

