\name{simspal}
\alias{simspal}
\docType{data}

\title{An example data set}

\description{
  Simulated data for an RIL data set. This data set have multiple QTL at
  1@61, 3@76 and 4@40.
}

\usage{data(simspal)}

\format{
  An object of class \code{cross}.  See \code{\link{read.cross}}
  for details.
}

\details{
  There are 162 RIL individuals typed at a total of 234 markers on five
  chromosomes.  The data were simulated with QTLs at 1@61, 3@76 and
  4@40.

}

\examples{
data(simspal)
}

\author{Il-Youp Kwak, \email{ikwak2@stat.wisc.edu} }

\keyword{datasets}