\docType{data}
\name{exd}
\alias{exd}
\title{An example data set}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(exd)
}
\description{
Simulated data for an backcross (with a QTL at 42 cM on chr
1) with a function-valued phenotype.
}
\examples{
data(exd)
}
\author{
Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\keyword{datasets}

