% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotlod.R
\name{plotlod}
\alias{plotlod}
\title{Plot signed or plain LOD image.}
\usage{
plotlod(output, effects, y, ylab = "Time", gap = 25, ncolors = 251,
  horizontal = FALSE, ...)
}
\arguments{
\item{output}{An object of class \code{"scanone"} as produced by
\code{\link[qtl]{scanone}}.}

\item{effects}{The sign information whether the QTL having AA do negative
effect.  Get this by using \code{\link{geteffects}}.}

\item{y}{Positions of phenotypes in image (e.g., the times).}

\item{ylab}{y-axis label}

\item{gap}{The gap between chromosomes.}

\item{ncolors}{The number of colors between blue and red.}

\item{horizontal}{If TRUE, transpose the x and y axes}

\item{\dots}{More graphical components, passed \code{\link[fields]{image.plot}}.}
}
\value{
None.
}
\description{
Plot signed or plain LOD image. (taking -LOD when the QTL effect
is negative), where (i,j) element is for phenotype i at marker j.
}
\examples{
data(simspal)
simspal <- calc.genoprob(simspal)
phe <- 1:nphe(simspal)
\dontshow{phe <- seq(1, nphe(simspal), by=60)}
out <- scanone(simspal, pheno.col=phe, method="hk")
eff <- geteffects(simspal, pheno.cols=phe)
nam <- phenames(simspal)
y <- as.numeric(substr(nam, 2, nchar(nam)))/60
\dontshow{y <- y[phe]}
plotlod(out, eff, y,  gap=15)
plotlod(out, y=y, gap=15, horizontal = TRUE)
}
\author{
Karl W Broman, Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\seealso{
\code{\link{geteffects}}
}
\keyword{hplot}

