% Generated by roxygen2 (4.0.1): do not edit by hand
\name{refineqtlF}
\alias{refineqtlF}
\title{Refine the positions of QTL for function valued trait}
\usage{
refineqtlF(cross, pheno.cols, usec = c("slod", "mlod"), qtl, covar = NULL,
  formula, method = c("hk", "imp"), verbose = TRUE, maxit = 10,
  incl.markers = TRUE, keeplodprofile = TRUE)
}
\arguments{
\item{cross}{An object of class \code{"cross"}. See \code{\link[qtl]{read.cross}} for details.}

\item{pheno.cols}{Columns in the phenotype matrix to be used as the
phenotype.}

\item{usec}{Which method to use (\code{"slod"} or \code{"mlod"})}

\item{qtl}{A QTL object, as produced by \code{\link[qtl]{makeqtl}}, containing the positions
of the QTL.}

\item{covar}{A matrix or data.frame of covariates.  These must be strictly
numeric.}

\item{formula}{An object of class \code{"formula"} indicating the model to be
fitted.  (It can also be the character string representation of a formula.)
QTLs are indicated as 'Q1', 'Q2', etc.  Covariates are indicated by their
names in \code{covar}.}

\item{method}{Indicates whether to use multiple imputation or Haley-Knott
regression.}

\item{verbose}{If TRUE, give feedback about progress.  If \code{verbose} is an
integer > 1, further messages from \code{\link[qtl]{scanqtl}} are also displayed.}

\item{maxit}{Maximum number of iterations.}

\item{incl.markers}{If FALSE, do calculations only at points on an evenly
spaced grid.}

\item{keeplodprofile}{If TRUE, keep the LOD profiles from the last iteration
as attributes to the output.}
}
\value{
An object of class \code{"qtl"}, with QTL placed in their new positions.
}
\description{
Extended version of the R/qtl function \code{\link[qtl]{refineqtl}} for
function-valued traits.  Iteratively scan the positions for QTL in the context of a
multiple QTL model, to try to identify the positions that maximize SLOD
criteria, for a fixed QTL model.
}
\examples{
data(exd)
exd <- calc.genoprob(exd, step=2)
qtl1.c <- makeqtl(exd, chr = 2, pos = 30, what = "prob")
thisqtl1.c <- refineqtlF(exd, pheno.cols = 1:10,  qtl = qtl1.c,
                         method = "hk")
}
\author{
Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\references{
Zeng, Z.-B., Kao, C.-H., and Basten, C. J. (1999) Estimating the
genetic architecture of quantitative traits.  _Genet. Res._ *74*, 279-289.

Haley, C. S. and Knott, S. A. (1992) A simple regression method for mapping
quantitative trait loci in line crosses using flanking markers.  _Heredity_
*69*, 315-324.
}
\seealso{
\code{\link[qtl]{refineqtl}}, \code{\link{refineqtlM}}
}
\keyword{models}

