% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scantwoF.R
\name{scantwoF}
\alias{scantwoF}
\title{Two-dimensional genome scan with a two-QTL model for function valued trait
data.}
\usage{
scantwoF(cross, pheno.cols, usec = c("slod", "mlod"), n.perm, ...)
}
\arguments{
\item{cross}{An object of class \code{"cross"}. See \code{\link[qtl]{read.cross}} for details.}

\item{pheno.cols}{Columns in the phenotype matrix to be used as the
phenotype.}

\item{usec}{Which method to use (\code{"slod"} or \code{"mlod"}) in the two-QTL scan.}

\item{n.perm}{If specified, a permutation test is performed rather than an
analysis of the observed data.  This argument defines the number of
permutation replicates.}

\item{\dots}{More parameters controled in 'scantwo'. See 'scantwo' for
details.}
}
\value{
If \code{n.perm} is missing, the function returns a list with class
\code{"scantwo"} and containing three components.  The first component is a
matrix of dimension [tot.pos x tot.pos]; the upper triangle contains the
(S/MLOD) scores for the additive model, and the lower triangle contains the
LOD scores for the full model.  The diagonal contains the results of
a single-QTL scan. The second component of the output is a data.frame indicating the
locations at which the two-QTL (S/MLOD) scores were calculated.  The first
column is the chromosome identifier, the second column is the position in
cM, the third column is a 1/0 indicator for ease in later pulling out only
the equally spaced positions, and the fourth column indicates whether the
position is on the X chromosome or not.  The final component is a version of
the results of 'scanone' including sex and/or cross direction as additive
covariates, which is needed for a proper calculation of conditional (S/MLOD)
scores.

If \code{n.perm} is specified, the function returns a list with six different LOD
scores from each of the permutation replicates. ... need more ..
}
\description{
Extension of the R/qtl function \code{\link[qtl]{scantwo}}. Two-dimensional genome
scan with a two-QTL model for function valuded trait data.
}
\examples{
data(exd)
exd <- calc.genoprob(exd, step = 0)
out <-scantwoF(exd, method = "hk", usec="slod")
plot(out)

# Permutation tests
n.perm <- 1000
\dontrun{permo <- scantwoF(exd, method="hk", n.perm=n.perm)
summary(permo, alpha=0.05)}
}
\seealso{
\code{\link[qtl]{scantwo}}, code{\link[funqtl]{scanoneF}}
}
\author{
Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\keyword{models}
