% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simspal-data.R
\docType{data}
\name{simspal}
\alias{simspal}
\title{An example data set}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(simspal)
}
\description{
Simulated data for an RIL data set. This data set have
multiple QTL on chromosomes 1 (61 cM), 3 (76 cM) and 4 (40 cM).
}
\examples{
data(simspal)
}
\author{
Il-Youp Kwak, <email: ikwak2@stat.wisc.edu>
}
\keyword{datasets}

